feat: 카카오 유저 탈퇴 기능 수정 (#60)

+++ api/src/main/java/com/bouquet/api/user/web/UserController.java
@@ -56,11 +56,11 @@ public class UserController {
                 hashMap.put("user", response);
                 hashMap.put("socialLoginType", 1);
                 hashMap.put("existingUser", "true");
-                try{
+                try {
                     hashMap.put("accessToken", jwtUtil.createToken("email", userInfo.getEmail()));
-                    hashMap.put("message", SUCCESS );
+                    hashMap.put("message", SUCCESS);
                     status = HttpStatus.ACCEPTED;
-                }catch(Exception e){
+                } catch (Exception e) {
                     hashMap.put("message", FAIL);
                     status = HttpStatus.INTERNAL_SERVER_ERROR;
                 }
@@ -70,6 +70,7 @@ public class UserController {
             else {
                 User user = new User();
                 user.setEmail(kakaoUserInfo.getEmail());
+                user.setRefreshToken(kakaoTokenInfo.getRefreshToken());
                 httpSession.setAttribute("user", user);
                 hashMap.put("existingUser", "false");
                 hashMap.put("socialLoginType", 1);
@@ -85,10 +86,11 @@ public class UserController {
 
     @ApiOperation(value = "닉네임 입력", notes = "nickname, email 값을 입력받아 유저 생성 후 jwt와 user 정보 반환")
     @PostMapping(value = "/login/user/nickname")
-    public ResponseEntity<Map<String, Object>> getUser(@RequestBody User user){
+    public ResponseEntity<Map<String, Object>> getUser(@RequestBody User user) {
+        System.out.println(user.toString());
         HttpStatus status = null;
         HashMap<String, Object> result = userService.create(user);
-        if(result.get("message").equals("success"))
+        if (result.get("message").equals("success"))
             status = HttpStatus.ACCEPTED;
         else status = status = HttpStatus.INTERNAL_SERVER_ERROR;
         return new ResponseEntity<Map<String, Object>>(result, status);
@@ -100,38 +102,51 @@ public class UserController {
 //
         HttpStatus status = null;
         User user = new User();
-        if(userRepository.existsByEmail(email) != null){
+        if (userRepository.existsByEmail(email) != null) {
             user = userRepository.existsByEmail(email);
-        }else{
+        } else {
             user.setEmail(email);
         }
-        System.out.println(email+"이메일");
-        System.out.println(user+"찾은유저정보");
+        System.out.println(email + "이메일");
+        System.out.println(user + "찾은유저정보");
         httpSession.setAttribute("user", user);
         HashMap<String, Object> result = userService.checknick(user);
-        if(result.get("existingUser").equals("false")){
-            result.put("email",email);
-            status = HttpStatus.ACCEPTED;}
-        else if(result.get("existingUser").equals("true") && result.get("message").equals("success")){
+        if (result.get("existingUser").equals("false")) {
+            result.put("email", email);
             status = HttpStatus.ACCEPTED;
-        }
-        else status = HttpStatus.INTERNAL_SERVER_ERROR;
+        } else if (result.get("existingUser").equals("true") && result.get("message").equals("success")) {
+            status = HttpStatus.ACCEPTED;
+        } else status = HttpStatus.INTERNAL_SERVER_ERROR;
         result.put("socialLoginType", 0);
         return new ResponseEntity<Map<String, Object>>(result, status);
     }
 
     @ApiOperation(value = "회원 탈퇴", notes = "userId 값을 입력받아 회원 탈퇴")
     @DeleteMapping(value = "/user/{userId}")
-    public ResponseEntity<UserResponse.OnlyId> delete(@PathVariable Long userId) {
-        UserResponse.OnlyId response = userService.delete(userId);
-        return ResponseEntity.ok().body(response);
+    public ResponseEntity<UserResponse.OnlyId> delete(@PathVariable Long userId, int socialLoginType) {
+        System.out.println("userId : " + userId + ", socialLoginType : " + socialLoginType);
+        //카카오 유저 탈퇴
+        if (socialLoginType == 1) {
+            //카카오 로그아웃
+            try {
+                String accessToken = kakaoAuthService.refreshToken(userId);
+                kakaoAuthService.kakaoUnlink(accessToken);
+                //kakaoAuthService.kakaoLogout(accessToken);
+                UserResponse.OnlyId response = userService.delete(userId);
+                return ResponseEntity.ok().body(response);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return ResponseEntity.internalServerError().build();
     }
+
     @ApiOperation(value = "리다이랙트 url", notes = "구글 로그인 시 url 뒤에 파라미터로 이메일을 넘김")
     @GetMapping(value = "/login/email")
     public void oauthLogin(HttpServletResponse response, HttpSession httpSession) throws IOException {
-        User user = (User)httpSession.getAttribute("user");
-        String redirect_uri="http://localhost:3000/login/redirect";
-        response.sendRedirect(redirect_uri+"?email="+user.getEmail());
+        User user = (User) httpSession.getAttribute("user");
+        String redirect_uri = "http://localhost:3000/login/redirect";
+        response.sendRedirect(redirect_uri + "?email=" + user.getEmail());
     }
